special TEID handling for GTPv2 Change Notification Req/Resp
GTPv2 needs special handling for TEID == 0 in Change Notification
Request and Responses. Earlier 3GPP releases specified that the
TEID in those messages should be 0, newer releases require a valid
TEID. In order to compatible with older versions, we need to handle
the TEID == 0 case.

Register all context with IMEI and IMSI, and use that to lookup
the context in those special cases.